편지 에디터 리팩토링
