+++ src/Components/paperStyles/PaperStyle.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {View} from 'react-native';
+import {Dotted} from './Dotted';
+import {Grid} from './Grid';
+
+interface Props {
+  lineColor: string;
+  paperStyle?: 'dotted' | 'grid' | 'plain';
+}
+
+export const PaperStyle = ({lineColor, paperStyle}: Props) => {
+  return (
+    <View style={{position: 'relative'}}>
+      {paperStyle === 'grid' && <Grid lineColor={lineColor} />}
+      {paperStyle === 'dotted' && <Dotted lineColor={lineColor} />}
+    </View>
+  );
+};
