+++ src/Constants/letter.ts
@@ -1,0 +1,267 @@
+import {Texticons} from '../types/types';
+
+export const PAPER_COLORS: string[] = [
+  '#fd40c1',
+  '#fe7542',
+  '#ffd94d',
+  '#bbfd50',
+  '#44febc',
+  '#41edfd',
+  '#3b7bfa',
+  '#806ef9',
+  '#dda0fb',
+];
+
+export const PAPER_STYLES = ['grid', 'dotted', 'plain'] as const;
+
+const HAPPY_TEXTICONS = [
+  '⸝⸝ʚ̴̶̷̆ᴗʚ̴̶̷̆⸝⸝',
+  '《 ͡꘠ ͜ゝ ͡꘠》',
+  '❣°ʚ(❛ั ᴗ ❛ั)ɞ°❣',
+  'ฅ(^▸ਉ◂^)ฅ',
+  '( ᵔᵒᵔ )ᵍᵒᵒᵈᵎᵎ',
+  '◌ ｡˚✩(  ›   ̫ ‹ )✩˚ ｡◌',
+  '(ꆧ⩊ꆧ )',
+  '( ˊᵕ̤ ̮ ᵕ̤ˋ )',
+  '❉𓂂(◕˓˳◕ )𓂂❉',
+  'ପ(｡ᵔ ⩊ ᵔ｡)ଓ',
+  '(｡⌒𖧉⌒)⋆.˚⊹⁺',
+  '( ◠ •̫̮ ◠)☆',
+  '(ˊ•͈ ˓ •͈ˋ)',
+  "(*' ꈊ'*)ゝ",
+  '(̂˃ꄃ˂ )̂',
+  '( • .̮ •)◞⸒⸒',
+  '（っ ‘ ᵕ ‘ ｃ）',
+  '(ᵔᵕᵔ)',
+  '●̑ᴗ●̑',
+  'ฅ(•ㅅ•❀)ฅ',
+  '₍๐•ᴗ•๐₎',
+  '-ˋˏ * ٩( ◡̉̈ )۶ * ˎˊ-',
+  '(｡•̀ᴗ-ღ)',
+  '◟(ᵔ ̮ ᵔ)͜💐',
+  '(〃´𓎟`〃)',
+  '(๑ ᴖ ༚̮ ᴖ ๑)',
+  'ᑫ•-³•ᑷ',
+];
+
+const ANGRY_TEXTICONS = [
+  '๛ก( ’-’* ก)',
+  '(ꐦ°᷄▿°᷅)',
+  'ʰᵘʰ (ꐦ○_○）✧',
+  'ᕙ( ︡’︡益’︠)ง',
+  '٩(`ω´٩ꐦ)',
+  '(✧˃̶͈̀ロ˂̶͈́)੭ु⁾⁾',
+  '█▬▬ ◟(`ﮧ´ ◟ )',
+  'ᕙ( ︡’︡ 益 ’︠)ง▬▬█',
+  '( ◡̀_◡́)▬▬█',
+  '₍ᐡ-᷅ ·̫ -᷄ᐡ₎',
+  '•̀ㅅ•́',
+  '(≡•̀·̯•́≡)',
+  '͡ ͜ʖ ͡ ╭∩╮',
+  'ฅʕ◍·̀·́◍ʔฅ',
+  '₍𝆥ﮦﮧﮦ₎𝆥',
+  '(҂⌣̀_⌣́)',
+  'ಠoಠ',
+  '(˘̶̀• ̯•˘̶́)',
+  '( ˘•~•˘ )',
+  '( `-´ )',
+  '\\( •̀ω•́ )//',
+  '(･⏠･)',
+  '( ‾᷅ᾥ‾᷄ )',
+  '٩(๑˃̌ۿ˂̌๑)۶',
+  '✧ `↼´˵',
+  '(-᷅_-᷄)',
+  '๛∙᷅῞ॄ∙᷄',
+  '(ꐦ°꒫°)',
+  'ᕕ(ꐦ°᷄д°᷅)ᕗ',
+];
+
+const WORRY_TEXTICONS = [
+  '･᷄-･᷅',
+  '(੭ ･᷄﹏･᷅)੭ु⁾',
+  'ㅇࡇㅇ?',
+  '(´╹〽╹`)',
+  '(ᐢ ܸ. .ܸ ᐢ)՞ ՞',
+  '(•́ι_•̀*)',
+  '=͟͟͞͞(๑º ﾛ º๑)',
+  '(˵ˊᯅˋ˵)',
+  '=͟͟͞͞(꒪ᗜ꒪ ‧̣̥̇)',
+  '( •︠ˍ•︡  )',
+  '(›´ω`‹ )',
+  '( ´△｀)',
+  '(;´・`)>',
+  '(･_-｡ )',
+  '(;´Д`)',
+  '（；￣ェ￣）',
+  '(;° ロ°)',
+  '(；☉_☉)',
+  '（’-’*)',
+  '(((( ;°Д°))))',
+  '((*゜Д゜)ゞ”',
+  '(」゜ロ゜)」',
+  '(´；Д；｀)',
+  '（￣□￣；）',
+  '(￣◇￣;)',
+  '(°◇°;)',
+  '(꒪⌓꒪)',
+  '（ｏ。ｏ；）',
+  '(ι´Д｀)ﾉ',
+];
+
+const UPFEELING_TEXTICONS = [
+  '(งᐛ)ว (งᐖ )ว',
+  '⸜(｡˃ ᵕ ˂ )⸝',
+  'ଘ꒰ ๑ ˃̶ ᴗ ᵒ̴̶̷๑꒱و  ̑̑',
+  '(｡˃ ᵕ ˂ )b',
+  '٩༼๑･ิﻌ･ิ༽۶٩༼･ิﻌ･ิ๑༽۶',
+  '٩(◕ᗜ◕)و',
+  '(๑˃̶͈̀∇˂̶͈́)و⁾⁾˚*',
+  'ꉂꉂ(ᵔᗜᵔ*)',
+  '(*•̀ᴗ•́*)و ̑̑',
+  '♡⁺◟(●˙▾˙●)◞⁺♡',
+  "(•'ٮ'•)৴",
+  '(ﾉ›_‹)ﾉ',
+  "٩'へ'و",
+  "٩(*'へ'*)و",
+  '(╯✧▽✧)╯',
+  'ʸᵉᵃʰ( ᐛ✌️)',
+  '٩( °ꇴ °)۶',
+  '(๑¯ ³¯๑)',
+  '੧ᐛ੭',
+  'ヽ(^。^)ノ',
+  '❛ ᗜ❛ ฅ',
+  '(੭•̀ᴗ•̀)੭',
+  '٩( ͡◉ ̯ ͡◉)۶',
+  "⸜(*'꒳'* )⸝",
+  '१✌˚◡˚✌५',
+  '( ´ ▽ ` )ﾉ',
+  '(,, ･∀･)ﾉ゛',
+  '(;-_-)ノ',
+  '(｡･ω･)ﾉﾞ',
+];
+
+const WINK_TEXTICONS = [
+  '(˵ •̀ ᴗ - ˵ ) ✧',
+  '(･ｪ-)',
+  '(－ｏ⌒)',
+  '(。^_・)ノ',
+  '（＾＿－）',
+  '（＾＿・）',
+  '（＾＿−）',
+  '（＾＿－）≡★',
+  '（○゜ε＾○）',
+  '(o‿∩)',
+  '๏[-ิ_•ิ]๏',
+  '~(＾◇^)/',
+  '☆～（ゝ。∂）',
+  '✩⃛˞(๑ꆨ৺ꉺ๑)',
+  '(๑ꄱͦॢ◡ु ˂̶͈́๑)',
+  '｡ﾟ+.ღ(ゝ◡ ⚈᷀᷁ღ)',
+  '(˃̶᷄︿๏）',
+  '(∗ᵒ̶̶̷̀ω˂̶́∗)੭₎₎̊₊♡',
+  '(*⸰‿-) (☉_☉) (-‿◦☀)',
+  'ヾ(*ゝω・*)ﾉ',
+  'ｄ(ゝc_,・●)',
+  'v(*’-^*)ｂ',
+  'v(°∇^*)⌒☆',
+  'σ(ﾟｰ^*)',
+  '۴(๑ꆨ◡ꉺ๑)',
+  'd(*ゝωб*)',
+  '(ﾟ∇^*)',
+  '৲( ᵒ ૩ᵕ )৴♡*৹',
+  '(๑ゝᴗම๑) ৷ਕკ~ෆ',
+];
+
+const SAD_TEXTICONS = [
+  '૮ ˃̣̣̥᷄ᴥ˂̣̣̥᷅ ა',
+  '˚‧º·(′̥̥̥ ゝ ‵̥̥̥)‧º·˚',
+  '˃̣̣̥᷄‸˂̣̣̥᷅',
+  '˚‧º·(˃̣̣̥˓̭˂̣̣̥)‧º·˚',
+  '( ´•̥̥̥o•̥̥̥`)',
+  '(´•̥̥̥‸•̥̥̥`)',
+  '(˘̩̩̩̩̩̩◅˘̩̩̩̩̩̩ )',
+  '( ꈨຶ ˙̫̮ ꈨຶ )',
+  '( ͒˃̩̩⌂˂̩̩ ͒)',
+  '(　-̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥᷄д-̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥̥᷅ )',
+  '(｡ᵕ̣̣̣̣̣̣ ꘍ ᵕ̣̣̣̣̣̣｡)',
+  '(´. ̯.`)',
+  '(o̴̶̷᷄ ̯o̴̶̷̥᷅)',
+  '(˘̩̩̩ ˆ ˘̩̩̩)',
+  '( ᵕ ̯ ᵕ̩̩ )',
+  '૮₍ o̴̶̷̥᷅.o̴̶̷᷄ ₎ა',
+  '(|||❛︵❛。)',
+  '(◞ ‸ ◟ㆀ)',
+  '(´• ͜. •̥`)✧',
+  '(´•̥ ᵔ •̥`)',
+  '( ˘•̥ _•̥ ˘ )',
+  '(｡ﹷ ‸ ﹷ ✿)',
+  'ˢᵒᵇ(ᵕ̣̣̣̣̣ ہ ᵕ̣̣̣̣̣̣ ✿)ˢᵒᵇ',
+  '｡°(°’-‘°)°｡',
+  '(৹ᵒ̴̶̷᷄ ฅᵒ̴̶̷᷅৹) ੭',
+  '(̂ ˃̥̥̥ ˑ̫ ˂̥̥̥ )̂',
+  '｡･ﾟ･(ﾟ`ω´ﾟ)･ﾟ･｡',
+  '( ੭ ›ꪴ .‹ꪴ )੭',
+  'દ ¨̯ﮦ૩',
+];
+
+const LOVE_TEXTICONS = [
+  '(ʃƪ ₀ ³₀)❥',
+  '(｡・‧̫・｡).*＊♡',
+  'ෆ{ @ᵔ֊̫ᵔ@ }ෆ',
+  '｡°(♥ω♥)｡',
+  '(⸝⸝o̴̶̷᷄ ·̭ o̴̶̷̥᷅⸝⸝)♡',
+  '(ˊσ̴̶̷̤⌄σ̴̶̷̤ˋ).♡*.',
+  '(⸝⸝- ̫ -⸝⸝)ෆ⃛',
+  '( つ◍’-’◍)╮—̳͟͞͞♡',
+  '( つ◍’-’◍)╮—̳͟͞͞♥',
+  '♡(ू•‧̫•ू⑅)',
+  "(っ' - ')╮=͟͟͞♡",
+  '( ᵅั ᴈ ᵅั;)',
+  "(~♡'ㅅ')~♡⁼³₌₃",
+  '(ﾉ´ з `)ノ—̳͟͞͞♡',
+  '.◜ɞ◝♡',
+  '♡( ˘ ･𐩅･˘)♡',
+  '( ˊ ᵕ ˋ )♡.°⑅',
+  '(*˙˘˙)♡',
+  '( ⸍ɞ̴̶̷ ·̫ ɞ̴̶̷⸌ )♡',
+  '(o´〰`o)♡*✲ﾟ*｡',
+  '(๓´˘`๓)♡',
+  'ᐡ´• ̫ •̥ ꜀ᐡ♡♡',
+  'ς( ᐡ-ﻌ•ᐡ ) ♡ (^•ﻌ•^ღ)',
+  '♡ᐡ・ ̫・`ᐡ♡',
+  'ღ꒡ ᴈ꒡)♡⃛(꒡ε ꒡ღ',
+  '♡( ૢ⁼̴̤̆ ꇴ ⁼̴̤̆ ૢ)~ෆ♡',
+  '♡(˃͈ દ ˂͈ ༶ )',
+  '♡˖(⁎ᐙॢ*)ॢ(*ॢᐕॢ⁎)♡॰ॱ',
+];
+
+export const TEXTICONS: Texticons = {
+  happy: {
+    key: '기쁨',
+    list: HAPPY_TEXTICONS,
+  },
+  angry: {
+    key: '화남',
+    list: ANGRY_TEXTICONS,
+  },
+  worry: {
+    key: '걱정',
+    list: WORRY_TEXTICONS,
+  },
+  upfeeling: {
+    key: '신남',
+    list: UPFEELING_TEXTICONS,
+  },
+  wink: {
+    key: '윙크',
+    list: WINK_TEXTICONS,
+  },
+  sad: {
+    key: '슬픔',
+    list: SAD_TEXTICONS,
+  },
+  love: {
+    key: '사랑',
+    list: LOVE_TEXTICONS,
+  },
+};
