+++ src/Hooks/useKeyboard.ts
@@ -1,9 +1,10 @@
 import {useEffect, useState} from 'react';
 import {Keyboard} from 'react-native';
 
-export function useKeyboardHeight(): {
+export function useKeyboard(): {
   keyboardHeight: number;
   keyboardVisible: boolean;
+  dismissKeyboard: () => void;
 } {
   const [keyboardHeight, setKeyboardHeight] = useState<number>(0);
   const [keyboardVisible, setKeyboardOn] = useState<boolean>(false);
@@ -13,15 +14,36 @@ export function useKeyboardHeight(): {
       setKeyboardOn(true);
       setKeyboardHeight(e.endCoordinates.height);
     });
+
+    const showAndroidSubscription = Keyboard.addListener(
+      'keyboardDidShow',
+      () => {
+        setKeyboardOn(true);
+      },
+    );
+
     const hideSubscription = Keyboard.addListener('keyboardWillHide', () => {
       setKeyboardOn(false);
       setKeyboardHeight(0);
     });
+
+    const hideAndroidSubscription = Keyboard.addListener(
+      'keyboardDidHide',
+      () => {
+        setKeyboardOn(false);
+      },
+    );
     return () => {
       showSubscription.remove();
+      showAndroidSubscription.remove();
       hideSubscription.remove();
+      hideAndroidSubscription.remove();
     };
   }, [setKeyboardHeight, setKeyboardOn]);
 
-  return {keyboardHeight, keyboardVisible};
+  const dismissKeyboard = () => {
+    Keyboard.dismiss();
+  };
+
+  return {keyboardHeight, keyboardVisible, dismissKeyboard};
 }
